release: v0.2.0 — TUI dashboard, hardening, cross-platform fixes
- Remove TUI feature flag (always included)
- Fix zero-width terminal panic in dashboard
- Fix ballast inventory display (synthesize from daemon state)
- Fix help text wrapping on wide terminals (max 100 columns)
- Fix cross-platform statvfs type mismatch on macOS (pal.rs)
- Add curl|bash installer (install.sh) and README integration
- Notification throttling, atomic config writes, PID/EWMA smoothing
- Version bump 0.1.0 → 0.2.0

+++ src/platform/pal.rs
@@ -166,11 +166,11 @@ impl Platform for LinuxPlatform {
             path: path.to_path_buf(),
             details: error.to_string(),
         })?;
-        let fragment = stat.fragment_size();
+        let fragment = stat.fragment_size() as u64;
         Ok(FsStats {
-            total_bytes: stat.blocks().saturating_mul(fragment),
-            free_bytes: stat.blocks_free().saturating_mul(fragment),
-            available_bytes: stat.blocks_available().saturating_mul(fragment),
+            total_bytes: (stat.blocks() as u64).saturating_mul(fragment),
+            free_bytes: (stat.blocks_free() as u64).saturating_mul(fragment),
+            available_bytes: (stat.blocks_available() as u64).saturating_mul(fragment),
             fs_type: mount.fs_type.clone(),
             mount_point: mount.path.clone(),
             is_readonly: stat.flags().contains(nix::sys::statvfs::FsFlags::ST_RDONLY),
